Make <select> <button>s have opening behavior by default
This patch makes <button>s in <select> with appearance:base-select have
the behavior which opens the listbox by default without the need for
type=popover. If there are multiple buttons and any of them have
type=popover, then only those one(s) will have the opening behavior.

This was resolved in openui here:
https://github.com/openui/open-ui/issues/939#issuecomment-1910837275

Bug: 1511354
Change-Id: I689e3f35f292586f83070a5443e382eedae1e555